exam/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
boto3==1.26.77
botocore==1.29.77
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
distlib==0.3.6
Django==4.1.3
django-redis==5.2.0
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
joblib==1.2.0
msgpack==1.0.4
numpy==1.23.5
Pillow==9.3.0
platformdirs==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
redis==4.4.0
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.14
virtualenv==20.19.0
zope.interface==5.5.2
